+++ ml-algorithms/src/main/java/org/opensearch/ml/engine/algorithms/sparse_encoding/SparseEncodingModel.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.ml.engine.algorithms.text_embedding;
+package org.opensearch.ml.engine.algorithms.sparse_encoding;
 
 import ai.djl.inference.Predictor;
 import ai.djl.modality.Input;
@@ -22,6 +22,7 @@ import org.opensearch.ml.common.output.model.ModelResultFilter;
 import org.opensearch.ml.common.output.model.ModelTensorOutput;
 import org.opensearch.ml.common.output.model.ModelTensors;
 import org.opensearch.ml.engine.algorithms.DLModel;
+import org.opensearch.ml.engine.algorithms.sparse_encoding.SparseEncodingTranslator;
 import org.opensearch.ml.engine.annotation.Function;
 
 import java.util.ArrayList;
@@ -29,13 +30,10 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static org.opensearch.ml.common.model.TextEmbeddingModelConfig.FrameworkType.SENTENCE_TRANSFORMERS;
-import static org.opensearch.ml.engine.ModelHelper.ONNX_ENGINE;
-import static org.opensearch.ml.engine.ModelHelper.PYTORCH_ENGINE;
-
 @Log4j2
-@Function(FunctionName.TOKENIZE)
-public class TokenizerModel extends DLModel {
+@Function(FunctionName.SPARSE_ENCODING)
+public class SparseEncodingModel extends DLModel {
+
     @Override
     public ModelTensorOutput predict(String modelId, MLInput mlInput) throws TranslateException {
         MLInputDataset inputDataSet = mlInput.getInputDataset();
@@ -53,25 +51,20 @@ public class TokenizerModel extends DLModel {
     }
 
     @Override
+    public Translator<Input, Output> getTranslator(String engine, MLModelConfig modelConfig) {
+        return new SparseEncodingTranslator();
+    }
+
+    @Override
     public TranslatorFactory getTranslatorFactory(String engine, MLModelConfig modelConfig) {
         return null;
     }
 
     @Override
-    public Translator<Input, Output> getTranslator(String engine, MLModelConfig modelConfig) {
-        return new TokenizerTranslator(null);
-    }
-
-
-    @Override
     public Map<String, Object> getArguments(MLModelConfig modelConfig) {
-        Map<String, Object> arguments = new HashMap<>();
-        if (modelConfig == null){
-            return arguments;
-        }
         TextEmbeddingModelConfig textEmbeddingModelConfig = (TextEmbeddingModelConfig) modelConfig;
         Integer modelMaxLength = textEmbeddingModelConfig.getModelMaxLength();
-
+        Map<String, Object> arguments = new HashMap<>();
         if (modelMaxLength != null) {
             arguments.put("modelMaxLength", modelMaxLength);
         }
@@ -81,12 +74,10 @@ public class TokenizerModel extends DLModel {
     @Override
     public void warmUp(Predictor predictor, String modelId, MLModelConfig modelConfig) throws TranslateException {
         TextEmbeddingModelConfig textEmbeddingModelConfig = (TextEmbeddingModelConfig) modelConfig;
+        Integer modelMaxLength = textEmbeddingModelConfig.getModelMaxLength();
         String warmUpSentence = "warm up sentence";
-        if (modelConfig  != null) {
-            Integer modelMaxLength = textEmbeddingModelConfig.getModelMaxLength();
-            if (modelMaxLength != null) {
-                warmUpSentence = "sentence ".repeat(modelMaxLength);
-            }
+        if (modelMaxLength != null) {
+            warmUpSentence = "sentence ".repeat(modelMaxLength);
         }
         // First request takes longer time. Predict once to warm up model.
         Input input = new Input();
